Continue on data type fixing

+++ general/base/cpp11addition.h
@@ -80,7 +80,7 @@ int vectorindexOf(const std::vector<T> &list,const T &item)
     if(r==list.cend())
         return -1;
     else
-        return std::distance(list.cbegin(),r);
+        return static_cast<int>(std::distance(list.cbegin(),r));
 }
 
 template <class T>
